refactor(app): move subcommands into commands/ module with layered bootstrap
- Relocate bridge, team_guide, team_stdio into src/commands/ directory
- Extract init_environment (logging + config) and init_data_layer (materialize + db) as reusable bootstrap layers
- Isolate HTTP server logic into commands/server.rs
- Remove pub mod exports from lib.rs (these are binary-private modules)
- Refactor idle_scanner to accept configurable scan interval and use seconds-based timeout
