+++ crates/aionui-app/src/commands/team_stdio.rs
