+++ crates/aionui-app/src/main.rs
@@ -1,17 +1,16 @@
 mod cli;
+mod commands;
 
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use std::process::ExitCode;
-use std::time::Instant;
 
 use anyhow::Result;
 use clap::Parser;
-use tokio::net::TcpListener;
-use tracing::info;
 use tracing_subscriber::{EnvFilter, Layer, fmt, layer::SubscriberExt, util::SubscriberInitExt};
 
-use aionui_app::{AppConfig, AppServices, bridge, create_router, guide_stdio, team_stdio};
+use aionui_app::AppServices;
 use cli::{Cli, Command};
+use commands::{init_data_layer, init_environment};
 
 const NOISE_SUPPRESSIONS: &[&str] = &["sqlx::query=warn", "hyper_util=warn", "reqwest=warn"];
 
@@ -121,178 +120,14 @@ fn main() -> Result<ExitCode> {
 async fn async_main(merged_path: String, cli: Cli) -> Result<ExitCode> {
     // MCP stdio helpers must not touch the database, logging setup, or `AppServices`.
     match cli.command {
-        Some(Command::McpBridge) => Ok(bridge::run_mcp_bridge().await),
-        Some(Command::McpGuideStdio) => Ok(guide_stdio::run_guide_stdio().await),
-        Some(Command::McpTeamStdio) => Ok(team_stdio::run_team_stdio().await),
-        None => run_server(cli, merged_path).await,
-    }
-}
-
-/// Resolve the conversation workspace directory.
-///
-/// Priority: `--work-dir` CLI flag → `AIONUI_WORK_DIR` env (when non-empty) →
-/// `--data-dir` fallback.
-fn resolve_work_dir(cli_work_dir: Option<PathBuf>, data_dir: &Path) -> PathBuf {
-    cli_work_dir.unwrap_or_else(|| {
-        std::env::var("AIONUI_WORK_DIR")
-            .ok()
-            .filter(|s| !s.is_empty())
-            .map(PathBuf::from)
-            .unwrap_or_else(|| data_dir.to_path_buf())
-    })
-}
-
-/// Materialize the embedded builtin-skills corpus to disk and clean up
-/// pre-symlink era stale directories.
-///
-/// Gated by a `.version` file so this is a no-op on subsequent starts with
-/// the same binary. When `AIONUI_BUILTIN_SKILLS_PATH` is set, skip
-/// materialization — the override path is the source of truth in that mode.
-async fn materialize_builtin_skills(data_dir: &Path) -> Result<()> {
-    let skip = std::env::var(aionui_extension::BUILTIN_SKILLS_ENV_VAR)
-        .map(|v| !v.is_empty())
-        .unwrap_or(false);
-    if skip {
-        return Ok(());
-    }
-
-    aionui_extension::materialize_if_needed(
-        data_dir,
-        aionui_extension::builtin_skills_corpus(),
-        env!("CARGO_PKG_VERSION"),
-    )
-    .await
-    .map_err(|e| anyhow::anyhow!("Failed to materialize builtin skills: {e}"))?;
-
-    // Best-effort cleanup of directories left behind by pre-symlink
-    // refactors. Failures are non-fatal — stale empty dirs are harmless.
-    // Runs ONLY after `materialize_if_needed` succeeded so we never touch
-    // data_dir until the builtin skills tree is in place. Do NOT
-    // generalize this list — it is an explicit allowlist of known-dead
-    // directories.
-    for stale in ["builtin-skills-view", "tmp", "agent-skills"] {
-        let path = data_dir.join(stale);
-        if path.exists()
-            && let Err(e) = std::fs::remove_dir_all(&path)
-        {
-            tracing::warn!(
-                path = %path.display(),
-                error = %e,
-                "failed to clean up stale data dir entry",
-            );
-        }
-    }
-    Ok(())
-}
-
-async fn run_server(cli: Cli, merged_path: String) -> Result<ExitCode> {
-    let log_dir = cli.log_dir.unwrap_or_else(|| cli.data_dir.join("logs"));
-    let _log_guard = init_tracing(&log_dir, cli.log_level.as_deref());
-
-    tracing::info!(
-        path_segments = merged_path.split(if cfg!(windows) { ';' } else { ':' }).count(),
-        path_len = merged_path.len(),
-        "startup: PATH ready"
-    );
-
-    let work_dir = resolve_work_dir(cli.work_dir, &cli.data_dir);
-
-    // SAFETY: called before any service initialization; no concurrent reads.
-    unsafe {
-        std::env::set_var("AIONUI_WORK_DIR", &work_dir);
-    }
-
-    let config = AppConfig {
-        host: cli.host,
-        port: cli.port,
-        data_dir: cli.data_dir,
-        work_dir,
-        app_version: cli.app_version,
-        local: cli.local,
-    };
-    if config.local {
-        info!("Running in local mode — authentication is disabled");
-    }
-
-    let boot = Instant::now();
-
-    let db_path = config.database_path();
-    aionui_db::maybe_copy_legacy_database(&db_path)?;
-    info!("Initializing database at {}", db_path.display());
-    let database = aionui_db::init_database(&db_path).await?;
-    info!(elapsed_ms = boot.elapsed().as_millis(), "startup: database initialized");
-
-    materialize_builtin_skills(&config.data_dir).await?;
-    info!(elapsed_ms = boot.elapsed().as_millis(), "startup: builtin skills materialized");
-
-    let services = AppServices::from_config(database, &config).await?;
-    info!(elapsed_ms = boot.elapsed().as_millis(), "startup: services constructed");
-
-    let has_users = services.user_repo.has_users().await?;
-    if !has_users {
-        info!("No configured users detected — initial setup required via /api/auth/status");
-    }
-
-    let router = create_router(&services).await;
-    let addr = config.socket_addr();
-    let listener = TcpListener::bind(&addr).await?;
-
-    info!(elapsed_ms = boot.elapsed().as_millis(), "Server listening on {addr}");
-
-    // Kick off the idle-ACP-agent reaper. `start_idle_scanner` returns
-    // immediately with a `JoinHandle`; the scanner task polls every 60 s
-    // and kills ACP agents whose `status == Finished` + last_activity
-    // exceeds the default 5-minute idle threshold. The watch channel
-    // propagates graceful-shutdown so the scanner exits on SIGINT/SIGTERM.
-    let (shutdown_tx, shutdown_rx) = tokio::sync::watch::channel(false);
-    let idle_scanner_handle =
-        aionui_ai_agent::start_idle_scanner(services.worker_task_manager.clone(), None, shutdown_rx);
-
-    axum::serve(listener, router)
-        .with_graceful_shutdown(async move {
-            shutdown_signal().await;
-            // Best-effort: only fails if every receiver has been dropped, in
-            // which case the scanner is already gone and there is nothing to
-            // signal.
-            let _ = shutdown_tx.send(true);
-        })
-        .await?;
-
-    // Wait for the scanner to observe the shutdown watch value and
-    // return; at worst this blocks for the current 60 s tick. Log-only
-    // on failure since the HTTP server has already drained.
-    if let Err(e) = idle_scanner_handle.await {
-        tracing::warn!(error = %e, "idle scanner join failed");
-    }
-
-    services.database.close().await;
-    info!("Server shut down gracefully");
-
-    Ok(ExitCode::SUCCESS)
-}
-
-async fn shutdown_signal() {
-    let ctrl_c = async {
-        tokio::signal::ctrl_c().await.expect("failed to install Ctrl+C handler");
-    };
-
-    #[cfg(unix)]
-    let terminate = async {
-        tokio::signal::unix::signal(tokio::signal::unix::SignalKind::terminate())
-            .expect("failed to install SIGTERM handler")
-            .recv()
-            .await;
-    };
-
-    #[cfg(not(unix))]
-    let terminate = std::future::pending::<()>();
-
-    tokio::select! {
-        () = ctrl_c => {
-            info!("Received SIGINT, shutting down...");
-        }
-        () = terminate => {
-            info!("Received SIGTERM, shutting down...");
+        Some(Command::McpBridge) => Ok(commands::run_mcp_bridge().await),
+        Some(Command::McpGuideStdio) => Ok(commands::run_team_guide().await),
+        Some(Command::McpTeamStdio) => Ok(commands::run_team_stdio().await),
+        None => {
+            let env = init_environment(&cli, &merged_path)?;
+            let database = init_data_layer(&env.config).await?;
+            let services = AppServices::from_config(database, &env.config).await?;
+            commands::run_server(env, services).await
         }
     }
 }
